MG improved heur.c, using time limit and implementing a better plotting

+++ src/tsp.h
@@ -44,7 +44,7 @@ typedef struct {
 	double* costs;
 	
 /*parameters*/
-	long timelimit; /*total time limit*/
+	double timelimit; /*total time limit*/
 	int randseed;
 	int n_sim;
 	int prob; /*probability*/
